log_service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
pika==1.3.2
mongoengine==0.27.0
pymongo==4.6.0